Image analysis code
This is a cleaned-up version of bwf.py. After a lot of testing, I realized that the only conditions you need to determine the beam width at any point are fitting a gaussian (for a gaussian or super-gaussian beam) or a hard edge circle (for a geometric beam), in which case you just need to find the widest part of the beam using a thresholding value.  In addition, if you uncomment both methods and line 246 (comment 247), you can graph both the gaussian and hard edge and compare with just one run. I personally found this helpful for trying to see which technique provided a more desirable result. 
Finally, there's a section (lines 291-304 & 349-354) that can fit a line so that one can compare the different results from fitting a Rayleigh range beam width eqn or from fitting a linear beam width. The line fits based on points that have a threshold of the maximum slope. This lets us fit the line based on what parts of the beam are roughly liner, such as far away from the beam waist